// src/modules/scrape/aws-imd/types/IProcessFileResult.ts
export type IObservation = {
  uid: string;
  station: string;
  date: string;
  time_utc: string;
  rainfall_mm: string;
  temperature_celcius: string;
  relative_humidity: string;
  wind_dir_10m_degree: string;
  wind_speed_10m_knots: string;
  max_wind_speed_by_gust_10m_knots: string;
  wind_dir_3m_degree: string;
  wind_speed_3m_knots: string;
  max_wind_speed_by_gust_3m_knots: string;
  sea_level_pressure_hpa: string;
  mean_sea_level_pressure: string;
  sunshine_hh_mm: string;
  soil_temp_at_10cm_depth_celcius: string;
  soil_moisture_at_10cm_depth: string;
  soil_temp_at_30cm_depth_celcius: string;
  soil_moisture_at_30cm_depth: string;
  soil_temp_at_70cm_depth_celcius: string;
  soil_moisture_at_70cm_depth: string;
  soil_temp_at_100cm_depth_celcius: string;
  soil_moisture_at_100cm_depth: string;
  global_radiation: string;
  photo_active_radiation: string;
  cat_code: string;
};